+++ packages/mx-adapter/playwright/mx.playwright.ts
@@ -18,7 +18,7 @@ test("connects to mx bank with oAuth, tracks performance correctly, and does ref
   const accessToken = await getAccessToken(request);
 
   await page.goto(
-    `http://localhost:8080/widget?jobTypes=${ComboJobTypes.TRANSACTIONS}&userId=${userId}`,
+    `http://localhost:8080/widget?jobTypes=${ComboJobTypes.TRANSACTIONS}&userId=${userId}&targetOrigin=http://localhost:8080`,
   );
 
   page.evaluate(`
@@ -84,15 +84,16 @@ test("connects to mx bank with oAuth, tracks performance correctly, and does ref
       const obj = (await msg.args()[0].jsonValue())?.message;
       if (obj?.type === "connect/memberConnected") {
         clearTimeout(timer);
-        expect(obj.metadata.user_guid).not.toBeNull();
-        expect(obj.metadata.member_guid).not.toBeNull();
-        expect(obj.metadata.aggregator).toEqual("mx_int");
+        expect(obj.metadata.user_guid).toBeUndefined();
+        expect(obj.metadata.aggregatorUserId).not.toBeNull();
         expect(obj.metadata.connectionId).not.toBeNull();
+        expect(obj.metadata.member_guid).toBeUndefined();
+        expect(obj.metadata.aggregator).toEqual("mx_int");
 
-        const { member_guid, aggregator, ucpInstitutionId } = obj.metadata;
+        const { connectionId, aggregator, ucpInstitutionId } = obj.metadata;
 
         await page.goto(
-          `http://localhost:8080/widget?jobTypes=${ComboJobTypes.TRANSACTIONS}&userId=${userId}&aggregator=${aggregator}&institutionId=${ucpInstitutionId}&connectionId=${member_guid}`,
+          `http://localhost:8080/widget?jobTypes=${ComboJobTypes.TRANSACTIONS}&userId=${userId}&aggregator=${aggregator}&institutionId=${ucpInstitutionId}&connectionId=${connectionId}&targetOrigin=http://localhost:8080`,
         );
 
         const popupPromise2 = page.waitForEvent("popup");
@@ -136,7 +137,7 @@ test("results in a successful performance event even if you close the tab", asyn
   const accessToken = await getAccessToken(request);
 
   await page.goto(
-    `http://localhost:8080/widget?jobTypes=${ComboJobTypes.TRANSACTIONS}&userId=${userId}`,
+    `http://localhost:8080/widget?jobTypes=${ComboJobTypes.TRANSACTIONS}&userId=${userId}&targetOrigin=http://localhost:8080`,
   );
 
   page.evaluate(`
@@ -216,7 +217,7 @@ test("shows an error page if you deny an mx bank oauth connection", async ({
   const userId = crypto.randomUUID();
 
   await page.goto(
-    `http://localhost:8080/widget?jobTypes=${ComboJobTypes.TRANSACTIONS}&userId=${userId}`,
+    `http://localhost:8080/widget?jobTypes=${ComboJobTypes.TRANSACTIONS}&userId=${userId}&targetOrigin=http://localhost:8080`,
   );
 
   await page.getByPlaceholder("Search").fill("MX Bank (Oauth)");